refactor: project restracture
