+++ src/app/pages/catalog/components/quiz-card/quiz-card-colors.enums.ts
