+++ src/app/pages/catalog/components/quiz-card/quiz-card.component.html
