+++ src/app/pages/catalog/components/quiz-card/quiz-card.component.ts
@@ -1,18 +1,18 @@
 import { Component, input, OnInit } from '@angular/core';
-import { QUIZ_CARD_STYLES } from './ui-quiz-card.styles';
-import { QuizCardColors } from '../../enums/quiz-card-colors.enums';
-import { QuizCardStyle } from './ui-quiz-card.types';
-import { QuizCategory } from '../../models/quiz-category.model';
-import { DEFAULT_CARD_COLOR } from './ui-quiz-card.constants';
+import { QUIZ_CARD_STYLES } from './quiz-card.styles';
+import { QuizCardColors } from './quiz-card-colors.enums';
+import { QuizCardStyle } from './quiz-card.types';
+import { QuizCategory } from '../../../../shared/models/quiz-category.model';
+import { DEFAULT_CARD_COLOR } from './quiz-card.constants';
 import { RouterLink } from '@angular/router';
 
 @Component({
   selector: 'quiz-quiz-card',
   standalone: true,
   imports: [RouterLink],
-  templateUrl: './ui-quiz-card.component.html',
+  templateUrl: './quiz-card.component.html',
 })
-export class UiQuizCardComponent implements OnInit {
+export class QuizCardComponent implements OnInit {
   quizCategory = input<QuizCategory>();
 
   categoryUrlRoute!: string[];
