+++ src/app/pages/quiz/components/question-card/question-card.component.html
