+++ src/app/pages/quiz/quiz.component.html
@@ -12,7 +12,7 @@
     Can you decipher these emoji clues to answer these music-related questions?
   </p>
 
-  <quiz-ui-question-card></quiz-ui-question-card>
+  <quiz-question-card></quiz-question-card>
 
   <quiz-ui-button class="my-4" variant="accent" route="/catalog"
     >Back to all Quizzes</quiz-ui-button
