+++ src/app/pages/quiz/quiz.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { UiQuestionCardComponent } from '../../shared/ui-kit/ui-question-card/ui-question-card.component';
+import { QuestionCardComponent } from './components/question-card/question-card.component';
 import { UiButtonComponent } from '../../shared/ui-kit/ui-button/ui-button.component';
 import { NavigationConfirmDialogComponent } from '../../shared/ui-kit/ui-navigation-confirm-dialog/ui-navigation-confirm-dialog.component';
 
@@ -7,7 +7,7 @@ import { NavigationConfirmDialogComponent } from '../../shared/ui-kit/ui-navigat
   selector: 'quiz-quiz',
   standalone: true,
   imports: [
-    UiQuestionCardComponent,
+    QuestionCardComponent,
     UiButtonComponent,
     NavigationConfirmDialogComponent,
   ],
