+++ src/app/services/randomization-service/randomization.service.ts
@@ -1,4 +1,4 @@
-import { QuizCardColors } from '../../shared/enums/quiz-card-colors.enums';
+import { QuizCardColors } from '../../pages/catalog/components/quiz-card/quiz-card-colors.enums';
 import { Injectable } from '@angular/core';
 
 @Injectable({
